# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Misko API
  version: 0.2.0
  description: >-
    Health, authentication, user management and the laboratory singleton.
    Every error response carries a stable `code` and an English `error` message.
    401 means missing, invalid, expired or revoked credentials; 403 means the
    current role lacks the permission. Roles are read from the database on every
    request, never from the token.
security:
  - bearer: []
paths:
...
  /api/paradigms:
    get:
      operationId: listParadigms
      summary: Paradigms with their published versions (*:read); the catalog is read-only
      responses:
        '200':
          description: Paradigms ordered by key
          content:
            application/json:
              schema: {type: object, required: [data], properties: {data: {type: array, items: {$ref: '#/components/schemas/ParadigmSummary'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
  /api/paradigms/{key}:
    parameters:
      - $ref: '#/components/parameters/ParadigmKey'
    get:
      operationId: getLatestParadigm
      summary: Manifest of the latest published version (*:read)
      responses:
        '200': {description: Manifest, content: {application/json: {schema: {$ref: '#/components/schemas/ParadigmManifest'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/paradigms/{key}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/ParadigmKey'
      - {name: version, in: path, required: true, schema: {type: integer, minimum: 1}}
    get:
      operationId: getParadigmVersion
      summary: Manifest of one published version, as consumed by analysis workers (*:read)
      responses:
        '200': {description: Manifest, content: {application/json: {schema: {$ref: '#/components/schemas/ParadigmManifest'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /api/tests/{id}/recordings:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      operationId: listRecordings
      summary: Recordings of a test with their video status (*:read)
      responses:
        '200':
          description: Recordings
          content:
            application/json:
              schema: {type: object, required: [data], properties: {data: {type: array, items: {$ref: '#/components/schemas/Recording'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    post:
      operationId: startRecordingUpload
      summary: Declare an original video and get a signed request that starts a resumable upload straight to Cloud Storage (test:run)
      description: Send upload.method to upload.url with upload.headers to get the resumable session URI from Cloud Storage, upload the file to it, then call finalize. The signed request is returned once and never stored.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RecordingUploadInput'}}}}
      responses:
        '201':
          description: Upload started
          content:
            application/json:
              schema:
                type: object
                required: [recording, upload]
                properties:
                  recording: {$ref: '#/components/schemas/Recording'}
                  upload: {$ref: '#/components/schemas/SignedUpload'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /api/tests/{id}/recordings/{recordingId}/finalize:
    parameters:
      - $ref: '#/components/parameters/ID'
      - {name: recordingId, in: path, required: true, schema: {type: string, format: uuid}}
    post:
      operationId: finalizeRecordingUpload
      summary: Verify the stored object and record the result (test:run; the uploader, LAB_MANAGER or SUPERADMIN)
      description: Size, CRC32C and content type are compared with the declaration. A mismatch stores status REJECTED with a reason. Finalizing again returns the stored result.
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses:
        '200': {description: Verified or rejected, content: {application/json: {schema: {$ref: '#/components/schemas/Recording'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
  /api/tests/{id}/recordings/{recordingId}/read-url:
    parameters:
      - $ref: '#/components/parameters/ID'
      - {name: recordingId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getRecordingReadUrl
      summary: A short-lived signed URL for the verified video generation (*:read); supports HTTP range requests
      responses:
        '200':
          description: Signed read URL
          content:
            application/json:
              schema: {type: object, required: [url, expiresAt], properties: {url: {type: string, format: uri}, expiresAt: {type: string, format: date-time}}}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '503': {$ref: '#/components/responses/Problem'}
...
components:
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    ParadigmKey: {name: key, in: path, required: true, schema: {type: string, pattern: '^[A-Z_]+$'}, example: OPEN_FIELD}
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 token with issuer, audience and expiry; see backend/README.md
    worker:
      type: apiKey
      in: header
      name: Authorization
      description: '"Worker TOKEN" with the token returned when the worker was registered; accepted only on /api/worker routes'
  responses:
    Problem:
      description: Error with a stable code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    ParadigmSummary:
      type: object
      required: [key, name, category, species, versions, latestVersion, automatedAnalysis]
      properties:
        key: {type: string}
        name: {type: string}
        category: {type: string, enum: [LEARNING_MEMORY, ANXIETY, MOTOR, SOCIAL]}
        species: {type: array, items: {type: string}}
        versions: {type: array, items: {type: integer}}
        latestVersion: {type: integer}
        automatedAnalysis: {type: boolean, description: Whether a real analysis worker supports this paradigm; catalog presence alone does not}
    ParadigmParameter:
      type: object
      required: [key, label, unit, valueType, min, max, default]
      properties:
        key: {type: string}
        label: {type: string}
        unit: {type: string, enum: [cm, cm_s, s, count, ratio, rpm, boolean]}
        valueType: {type: string, enum: [number, integer], description: Integer parameters accept whole numbers only; booleans are 0 or 1}
        min: {type: number}
        max: {type: number}
        default: {type: number}
    ParadigmMetric:
      type: object
      required: [key, label, unit, valueType, definition, formula, missingWhen, inputs, min, max, tolerance]
      properties:
        key: {type: string}
        label: {type: string}
        unit: {type: string, enum: [cm, cm_s, s, count, ratio, rpm, boolean]}
        valueType: {type: string, enum: [number, integer]}
        definition: {type: string}
        formula: {type: string}
        missingWhen: {type: string, description: When the metric is reported missing instead of a value}
        inputs: {type: array, items: {type: string}}
        min: {type: number}
        max: {type: number}
        tolerance: {type: number, description: Absolute tolerance in the metric unit}
    ParadigmEvent:
      type: object
      required: [type, kind, labels, single, definition]
      properties:
        type: {type: string}
        kind: {type: string, enum: [INTERVAL, POINT]}
        labels: {type: array, items: {type: string}, description: Allowed labels; empty when the event takes no label}
        single: {type: boolean, description: The event may occur at most once per trial}
        definition: {type: string}
    ParadigmManifest:
      type: object
      required: [key, name, category, species, trialTypes, version, metricEngineVersion, resultSchemaVersion, automatedAnalysis, apparatusParameters, sessionParameters, zones, inputEvents, events, metrics, qc]
      properties:
        key: {type: string}
        name: {type: string}
        category: {type: string}
        species: {type: array, items: {type: string}}
        trialTypes: {type: array, items: {type: string}}
        version: {type: integer, description: ParadigmVersion; a published version never changes}
        metricEngineVersion: {type: integer}
        resultSchemaVersion: {type: integer}
        automatedAnalysis: {type: boolean}
        apparatusParameters: {type: array, items: {$ref: '#/components/schemas/ParadigmParameter'}}
        sessionParameters: {type: array, items: {$ref: '#/components/schemas/ParadigmParameter'}}
        zones:
          type: array
          items:
            type: object
            required: [key, role, geometry, calibrated, required]
            properties:
              key: {type: string}
              role: {type: string}
              geometry: {type: string}
              calibrated: {type: boolean, description: Shapes come with the analysis input (per-video calibration) instead of being derived from parameters}
              required: {type: boolean}
        inputEvents:
          type: array
          description: Scored observations the metric engine consumes
          items: {$ref: '#/components/schemas/ParadigmEvent'}
        events:
          type: array
          description: Intervals and points the metric engine derives
          items: {$ref: '#/components/schemas/ParadigmEvent'}
        metrics: {type: array, items: {$ref: '#/components/schemas/ParadigmMetric'}}
        qc:
          type: array
          items: {type: object, required: [key, operator, value, unit], properties: {key: {type: string}, operator: {type: string}, value: {type: number}, unit: {type: string}}}
...
    Video:
      type: object
      required: [id, kind, contentType, fileName, sizeBytes, crc32c, status, generation, rejectionReason, verifiedAt, createdBy, createdAt]
      description: A video object in private Cloud Storage. Verified videos never change.
      properties:
        id: {type: string, format: uuid}
        kind: {type: string, enum: [ORIGINAL, ANALYZED]}
        contentType: {type: string, enum: [video/mp4, video/quicktime, video/webm]}
        fileName: {type: [string, 'null']}
        sizeBytes: {type: integer, format: int64}
        crc32c: {type: string, description: Base64 of the big-endian CRC32C}
        status: {type: string, enum: [PENDING, VERIFIED, REJECTED]}
        generation: {type: [string, 'null'], description: Cloud Storage object generation as a decimal string}
        rejectionReason: {type: [string, 'null'], enum: [OVERSIZED, SIZE_MISMATCH, CHECKSUM_MISMATCH, CONTENT_TYPE_MISMATCH, null]}
        verifiedAt: {type: [string, 'null'], format: date-time}
        createdBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    Recording:
      type: object
      required: [id, experimentId, testId, clipStartUs, clipEndUs, video, createdBy, createdAt]
      properties:
        id: {type: string, format: uuid}
        experimentId: {type: string, format: uuid}
        testId: {type: string, format: uuid}
        clipStartUs: {type: integer, format: int64, minimum: 0, description: Microseconds from the start of the video}
        clipEndUs: {type: [integer, 'null'], format: int64, description: Null runs to the end of the video}
        video: {$ref: '#/components/schemas/Video'}
        createdBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    RecordingUploadInput:
      type: object
      additionalProperties: false
      required: [fileName, contentType, sizeBytes, crc32c]
      properties:
        fileName: {type: string, minLength: 1, maxLength: 255}
        contentType: {type: string, enum: [video/mp4, video/quicktime, video/webm]}
        sizeBytes: {type: integer, format: int64, minimum: 1, description: At most VIDEO_MAX_BYTES}
        crc32c: {type: string, description: Base64 of the big-endian CRC32C of the whole file}
        clipStartUs: {type: integer, format: int64, minimum: 0, default: 0}
        clipEndUs: {type: integer, format: int64}
    SignedUpload:
      type: object
      required: [method, url, headers, expiresAt]
      description: A signed request that starts a resumable upload; it fails if the object already exists
      properties:
        method: {type: string, enum: [POST]}
        url: {type: string, format: uri}
        headers: {type: object, additionalProperties: {type: string}}
        expiresAt: {type: string, format: date-time}
...
    Problem:
      type: object
      additionalProperties: false
      required: [error, code]
      properties:
        error: {type: string}
        code: {type: string}
...